requirements.txt
pyspark==3.3.0
requests==2.28.1
hdfs==2.7.0
numpy==1.23.3
pandas==1.5.0
python-dotenv==1.0.0
matplotlib==3.6.0
seaborn==0.12.0
plotly==5.11.0
kaleido==0.2.1
